collibra/requirements.txt
# Collibra Metadata Synchronization Dependencies
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0

